feat: initial ERC1155 handling rework

+++ contracts/facets/L1/AssetHandler/AssetHandler.sol
@@ -90,7 +90,7 @@ contract L1AssetHandler is IL1AssetHandler, SolidStateLayerZeroClient {
         address collection,
         uint256[] calldata tokenIds
     ) external payable {
-        for (uint i = 0; i < tokenIds.length; i++) {
+        for (uint256 i = 0; i < tokenIds.length; i++) {
             IERC721(collection).safeTransferFrom(
                 msg.sender,
                 address(this),
@@ -147,7 +147,7 @@ contract L1AssetHandler is IL1AssetHandler, SolidStateLayerZeroClient {
             ) = abi.decode(data, (address, address, uint256[]));
 
             // Transfer the ERC721 assets to the sender
-            for (uint i = 0; i < tokenIds.length; i++) {
+            for (uint256 i = 0; i < tokenIds.length; i++) {
                 IERC721(collection).safeTransferFrom(
                     address(this),
                     sender,
